Added gems.
Gems are now included. At the moment each brick contains a gem behind
them, and when broken the gem is set free. Two things can happen when a
gem comes alive, either fall , or bobble. When it bobbles on the ball
can hit it. So far, no collision is added to the gems, yet. Plan on
adding collision for gem-ball and gem-paddle. Also, got to randomize
what bricks receive gems.